move server to app.js
