+++ Source/server/package.json
@@ -18,6 +18,7 @@
     "jsonwebtoken": "^9.0.0",
     "mongoose": "^7.0.2",
     "nodemon": "^2.0.21",
+    "supertest": "^6.3.3",
     "web3": "^1.8.2"
   }
 }
